pyproject.toml
[project]
name = "rdw"
version = "0.1.0"
description = "Self Driving Challenge perception algorithms"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.3.4",
    "onnxruntime>=1.23.2",
    "opencv-python>=4.11.0.86",
    "pyyaml>=6.0.3",
    "typer>=0.20.0",
]


[project.scripts]
rdw-classical = "classical_prior.cli:cli_entry"
rdw-motion    = "motion_fusion.cli:cli_entry"



[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.uv]
package = true

[tool.setuptools.packages.find]
where = ["src"]